Move redirect URLs into their own controller
This allows us to stop links to redirects being active in museum displays. Fixes #959
